update enigma_test by moving shift test methods to shift_test
