+++ test/enigma_test.rb
@@ -9,42 +9,8 @@ class EnigmaTest < Minitest::Test
     @enigma = Enigma.new
   end
 
-  def test_it_exists_and_attributes
+  def test_it_exists
     assert_instance_of Enigma, @enigma
-  end
-
-  def test_it_can_produce_random_key
-    assert_equal 5, @enigma.key.length
-  end
-
-  def test_it_can_produce_alphabet_array
-    expected = ("a".."z").to_a << " "
-
-    assert_equal expected, @enigma.alphabet_array
-  end
-
-  def test_it_can_produce_last_four
-    assert_equal "1025", @enigma.last_four("040895")
-  end
-
-  def test_it_can_set_a_shifts
-    a = @enigma.alphabet_array.rotate(3)
-    assert_equal a, @enigma.set_a_shift("02715", "040895")
-  end
-
-  def test_it_can_set_b_shifts
-    b = @enigma.alphabet_array.rotate(27)
-    assert_equal b, @enigma.set_b_shift("02715", "040895")
-  end
-
-  def test_it_can_set_c_shifts
-    c = @enigma.alphabet_array.rotate(73)
-    assert_equal c, @enigma.set_c_shift("02715", "040895")
-  end
-
-  def test_it_can_set_d_shifts
-    d = @enigma.alphabet_array.rotate(20)
-    assert_equal d, @enigma.set_d_shift("02715", "040895")
   end
 
   def test_it_can_set_encrypt_message
